I don't know why, I don't want to know why, I shouldn't have to wonder why, but for some reason Github is not tracking the file renames unless I do this awfulness.
